Update dogpile.cache to 0.6.1 Note: removed dogpile.core 0.4.1 as it is not used by dogpile.cache since 0.6.0
