+++ lib/dogpile/__init__.py
@@ -1,6 +1,4 @@
-# See http://peak.telecommunity.com/DevCenter/setuptools#namespace-packages
-try:
-    __import__('pkg_resources').declare_namespace(__name__)
-except ImportError:
-    from pkgutil import extend_path
-    __path__ = extend_path(__path__, __name__)
+__version__ = '0.6.1'
+
+from .lock import Lock  # noqa
+from .lock import NeedRegenerationException  # noqa
