+++ lib/dogpile/cache/__init__.py
@@ -1,3 +1,4 @@
-__version__ = '0.5.7'
-
 from .region import CacheRegion, register_backend, make_region  # noqa
+
+# backwards compat
+from .. import __version__  # noqa
