--- lib/dogpile/core/__init__.py
@@ -1,11 +1,0 @@
-from .dogpile import NeedRegenerationException, Lock
-from .nameregistry import NameRegistry
-from .readwrite_lock import ReadWriteMutex
-from .legacy import Dogpile, SyncReaderDogpile
-
-__all__ = [
-        'Dogpile', 'SyncReaderDogpile', 'NeedRegenerationException',
-        'NameRegistry', 'ReadWriteMutex', 'Lock']
-
-__version__ = '0.4.1'
-
