--- lib/dogpile/core/util.py
@@ -1,8 +1,0 @@
-import sys
-py3k = sys.version_info >= (3, 0)
-
-try:
-    import threading
-except ImportError:
-    import dummy_threading as threading
-
