+++ lib/dogpile/util/__init__.py
@@ -1,0 +1,4 @@
+from .nameregistry import NameRegistry  # noqa
+from .readwrite_lock import ReadWriteMutex  # noqa
+from .langhelpers import PluginLoader, memoized_property, \
+    coerce_string_conf, to_list, KeyReentrantMutex  # noqa
