+++ lib/dogpile/util/compat.py
